+++ CHANGELOG.md
@@ -1,11 +1,17 @@
+## [0.0.4] - SDK
+
+- Bump sdk
+- Added text style and build context extension helpers
+  - bold, semiBold, medium, normal, withColor, withScale, mq, textScalerScale
+
 ## [0.0.3] - SDK
 
-* Bump sdk
+- Bump sdk
 
 ## [0.0.2] - Addition
 
-* Added simpleUniq, copy, removeEmpty, removeFalsy, removeEmptyOrFalsy, without, notEq
+- Added simpleUniq, copy, removeEmpty, removeFalsy, removeEmptyOrFalsy, without, notEq
 
 ## [0.0.1] - Initial release
 
-* Initial release.
+- Initial release.